\name{lmclus-package}
\alias{lmclus-package}
\docType{package}
\title{
    Linear Manifold Clustering Clustering
}

\description{
    This package contains implementation of
    Linear Manifold Clustering Clustering method.
}

\details{

  Use \code{lmclus} to cluster dataset.
  Use \code{?lmclus.param} to see method prmeters.

}
\author{ Art Diky }

\keyword{ package }
\keyword{ clustering }

\examples{
packageDescription('lmclus')
system.file("packages", package="lmclus")
}
